pyproject.toml
[tool.poetry]
name = "async_api"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = [{include = "async_api"}]

[tool.poetry.dependencies]
python = "~3.10"
elasticsearch = {version = "7.9.1", extras = ["async"]}
fastapi = "0.90.0"
uvicorn = "0.19.0"
orjson = "3.8.7"
redis = {extras = ["hiredis"], version = "4.5.4"}
gunicorn = "20.1.0"
pydantic = {extras = ["dotenv"], version = "1.10.6"}
python-dotenv = "1.0.0"
backoff = "2.2.1"
pyjwt = "2.6.0"
httpx = "0.23.3"
cryptography = "40.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.2.1"
pytest-asyncio = "0.12.0"
httpx = "0.23.3"
click = "8.0.4"
respx = "^0.20.1"
# linters
black = "22.8.0"
flake8 = "6.0.0"
isort = "5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
   venv
)/
'''

[tool.isort]
line_length = 79
combine_as_imports = true
include_trailing_comma = true
lines_after_imports = 2
skip = ["venv",]
